backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.7.0
pydantic-settings>=2.1.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
pyyaml>=6.0.1
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.1.0
langchain-community>=0.2.0
python-dotenv>=1.0.0
alembic>=1.13.1
httpx>=0.26.0
websockets>=12.0
python-multipart>=0.0.6
